premiers exercices en C

+++ pointers_arrays_strings/2-strlen.c
@@ -12,7 +12,7 @@
 
 int _strlen(char *s)
 {
-int len = 0;
+long int len = 0;
 len = sizeof(s)+2;
 return len;
 }
